Split components for table entries into their own file, since they were getting quite large and cover a different topic than table details
