+++ src/containers/tabledetails.js
@@ -9,7 +9,8 @@ import {
 import {WideColumnWorkspace} from '../components/workspaces'
 import {FloatingActionButton, FloatingWorkspaceMenu} from '../components/floatingcontrols'
 import {TextAreaInputModal} from '../components/modals'
-import {TableDetailsSegment, TableEntriesSegment} from '../components/tabledetails'
+import {TableDetailsSegment} from '../components/tabledetails'
+import {TableEntriesSegment} from '../components/tableentries'
 import {TableEntryEditModal} from '../components/tableentry'
 import {addTableEntry, updateTableEntry} from '../actions/tabledetails'
 import {getTableId, getFullTableById} from '../selectors/tabledetails'
